gh-200. Move MessageQueue class to utils
