+++ apps/web-portal-client/lib/components/Button.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import className from 'classnames'
 
-import {childrenPropTypes} from '../utils/childrenPropTypes'
+import {childrenPropTypes} from '../utils'
 
 import './button.css'
